Move region constraint search into a separate module
This also inlines some of the region graph search methods
